+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Please add unreleased changes in the following style:
 PR Title ([#123](link to my pr))
 ```
 
+Update react-maplibre ([#34](https://github.com/maplibre/maplibre-react-native/issues/34))
+
 ## 9.1.0
 
 Update react to 18.2.0 and react-native to 0.72.1. ([#49](https://github.com/maplibre/maplibre-react-native/pull/49))